# specs/001-create-a-drag/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Teacher Aide Scheduler API
  version: 1.0.0
  description: REST API for drag-and-drop timetable scheduling system
servers:
  - url: http://localhost:5000/api
    description: Local development server
paths:
...
  /tasks:
    get:
      summary: List tasks
      tags: [Tasks]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [UNASSIGNED, ASSIGNED, IN_PROGRESS, COMPLETE]
        - name: category
          in: query
          schema:
            type: string
            enum: [PLAYGROUND, CLASS_SUPPORT, GROUP_SUPPORT, INDIVIDUAL_SUPPORT]
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create one-off task
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /recurring-tasks:
    post:
      summary: Create recurring task with RRULE
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/TaskCreate'
                - type: object
                  required: [recurrence_rule, expires_on]
                  properties:
                    recurrence_rule:
                      type: string
                      example: "FREQ=WEEKLY;BYDAY=MO,WE,FR"
                    expires_on:
                      type: string
                      format: date
      responses:
        '201':
          description: Recurring task created with assignments generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/Task'
                  assignments_created:
                    type: integer
  /tasks/{id}:
    get:
      summary: Get task by ID
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    put:
      summary: Update task (regenerates future assignments if recurring)
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                category:
                  type: string
                start_time:
                  type: string
                end_time:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/Task'
                  assignments_updated:
                    type: integer
    delete:
      summary: Delete task (future occurrences only)
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted
  
  # Assignments
...
  /requests:
    get:
      summary: List teacher requests
      tags: [Requests]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
      responses:
        '200':
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Request'
    post:
      summary: Submit teacher request
      tags: [Requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requesting_teacher, task_title, task_category, preferred_date, preferred_time]
              properties:
                requesting_teacher:
                  type: string
                task_title:
                  type: string
                task_category:
                  type: string
                  enum: [PLAYGROUND, CLASS_SUPPORT, GROUP_SUPPORT, INDIVIDUAL_SUPPORT]
                preferred_date:
                  type: string
                  format: date
                preferred_time:
                  type: string
                classroom_id:
                  type: integer
                  nullable: true
                notes:
                  type: string
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
  /requests/{id}:
    put:
      summary: Update request status (APPROVED creates task)
      tags: [Requests]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [APPROVED, REJECTED]
      responses:
        '200':
          description: Request updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  request:
                    $ref: '#/components/schemas/Request'
                  task_created:
                    $ref: '#/components/schemas/Task'
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [PLAYGROUND, CLASS_SUPPORT, GROUP_SUPPORT, INDIVIDUAL_SUPPORT]
        start_time:
          type: string
        end_time:
          type: string
        recurrence_rule:
          type: string
          nullable: true
        expires_on:
          type: string
          format: date
          nullable: true
        classroom_id:
          type: integer
          nullable: true
        notes:
          type: string
        status:
          type: string
          enum: [UNASSIGNED, ASSIGNED, IN_PROGRESS, COMPLETE]
        created_at:
          type: string
          format: date-time
    TaskCreate:
      type: object
      required: [title, category, start_time, end_time]
      properties:
        title:
          type: string
        category:
          type: string
          enum: [PLAYGROUND, CLASS_SUPPORT, GROUP_SUPPORT, INDIVIDUAL_SUPPORT]
        start_time:
          type: string
        end_time:
          type: string
        classroom_id:
          type: integer
        notes:
          type: string
...
    Request:
      type: object
      properties:
        id:
          type: integer
        requesting_teacher:
          type: string
        task_title:
          type: string
        task_category:
          type: string
        preferred_date:
          type: string
          format: date
        preferred_time:
          type: string
        classroom_id:
          type: integer
          nullable: true
        notes:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        created_at:
          type: string
          format: date-time
...